requirements.txt
pygame
scipy
e-pyquadtree
pygame_gui